requirements.txt
openai>=1.12.0
pytest==7.4.3
pytest-asyncio==0.21.1
pydantic==2.5.3
python-dotenv==1.0.0
aiohttp==3.9.1
redis==5.0.1
prometheus-client==0.19.0
pyyaml==6.0.1